Refactor redux actions, reducers, types into subdirs
