--- src/redux/actionTypes.js
@@ -1,14 +1,0 @@
-export const SET_REPOS_SELECTION = "SET_REPOS_SELECTION";
-export const SET_WORKSPACE_SELECTION = "SET_WORKSPACE_SELECTION";
-export const SET_REFRESH_MINS = "SET_REFRESH_MINS";
-export const SET_SHOULD_DATA_REFRESH = "SET_SHOULD_DATA_REFRESH";
-export const FETCH_WORKSPACES_BEGIN = "FETCH_WORKSPACES_BEGIN";
-export const FETCH_WORKSPACES_SUCCESS = "FETCH_WORKSPACES_SUCCESS";
-export const FETCH_WORKSPACES_FAILURE = "FETCH_WORKSPACES_FAILURE";
-export const SET_ACCESS_TOKEN = "SET_ACCESS_TOKEN";
-export const FETCH_REPOS_PAGES_BEGIN = "FETCH_REPOS_PAGES_BEGIN";
-export const FETCH_REPOS_PAGE_SUCCESS = "FETCH_REPOS_PAGE_SUCCESS";
-export const FETCH_REPOS_PAGE_FAILURE = "FETCH_REPOS_PAGE_FAILURE";
-export const FETCH_REPOS_PAGE_CANCELLED = "FETCH_REPOS_PAGE_CANCELLED";
-
-export const FETCH_REPOS_PAGES_END = "FETCH_REPOS_PAGES_END";
