+++ src/redux/external/actionTypes.js
@@ -1,0 +1,1 @@
+export const SET_ACCESS_TOKEN = "SET_ACCESS_TOKEN";
