+++ src/redux/refresh/actionTypes.js
@@ -1,0 +1,2 @@
+export const SET_REFRESH_MINS = "SET_REFRESH_MINS";
+export const SET_SHOULD_DATA_REFRESH = "SET_SHOULD_DATA_REFRESH";
