+++ src/redux/refresh/actions.js
@@ -1,0 +1,19 @@
+import {
+    SET_REFRESH_MINS,
+    SET_SHOULD_DATA_REFRESH
+} from './actionTypes';
+
+
+export const setRefreshMins = (mins) => ({
+    type: SET_REFRESH_MINS,
+    payload: {
+        mins
+    }
+});
+
+export const setShouldDataRefresh = (yesNo) => ({
+    type: SET_SHOULD_DATA_REFRESH,
+    payload: {
+        yesNo
+    }
+});
