+++ src/redux/repos/actionTypes.js
@@ -1,0 +1,6 @@
+export const SET_REPOS_SELECTION = "SET_REPOS_SELECTION";
+export const FETCH_REPOS_PAGES_BEGIN = "FETCH_REPOS_PAGES_BEGIN";
+export const FETCH_REPOS_PAGE_SUCCESS = "FETCH_REPOS_PAGE_SUCCESS";
+export const FETCH_REPOS_PAGE_FAILURE = "FETCH_REPOS_PAGE_FAILURE";
+export const FETCH_REPOS_PAGE_CANCELLED = "FETCH_REPOS_PAGE_CANCELLED";
+export const FETCH_REPOS_PAGES_END = "FETCH_REPOS_PAGES_END";
