+++ src/redux/workspaces/actionTypes.js
@@ -1,0 +1,4 @@
+export const SET_WORKSPACE_SELECTION = "SET_WORKSPACE_SELECTION";
+export const FETCH_WORKSPACES_BEGIN = "FETCH_WORKSPACES_BEGIN";
+export const FETCH_WORKSPACES_SUCCESS = "FETCH_WORKSPACES_SUCCESS";
+export const FETCH_WORKSPACES_FAILURE = "FETCH_WORKSPACES_FAILURE";
